+++ apis/_api/index.ts
@@ -2,4 +2,3 @@ export * as authApi from './auth';
 export * as systemApi from './system';
 export * as clothApi from './cloth';
 export * as ootdApi from './ootd';
-export * as bookmarkApi from './bookmark';
